+++ web-ui/src/components/common/blob-visuals/blob-parts/tools/Dumbbells.tsx
@@ -1,0 +1,61 @@
+export const Dumbbells = () => {
+  return (
+    <>
+      <rect
+        fill="#b3b3b3"
+        stroke="#000000"
+        strokeWidth="1.0"
+        id="rect-rod"
+        width="6"
+        height="32"
+        x="100"
+        y="53"
+        transform="rotate(45 103 69)"
+      />
+      <rect
+        fill="#b3b3b3"
+        stroke="#000000"
+        strokeWidth="1.0"
+        id="rect-plate-u1"
+        width="6"
+        height="32"
+        x="116"
+        y="53"
+        transform="rotate(-45 103 69)"
+      />
+      <rect
+        fill="#b3b3b3"
+        stroke="#000000"
+        strokeWidth="1.0"
+        id="rect-plate-u2"
+        width="6"
+        height="24"
+        x="122"
+        y="57"
+        transform="rotate(-45 103 69)"
+      />
+      <rect
+        fill="#b3b3b3"
+        stroke="#000000"
+        strokeWidth="1.0"
+        id="rect-plate-l2"
+        width="6"
+        height="24"
+        x="122"
+        y="57"
+        transform="rotate(135 103 69)"
+      />
+      <rect
+        fill="#b3b3b3"
+        stroke="#000000"
+        strokeWidth="1.0"
+        id="rect-plate-l1"
+        width="6"
+        height="32"
+        x="116"
+        y="53"
+        transform="rotate(135 103 69)"
+      />
+    </>
+  );
+};