+++ web-ui/src/components/common/blob-visuals/blob-parts/tools/Hammer.tsx
@@ -1,0 +1,25 @@
+export const Hammer = () => {
+  return (
+    <>
+      <rect
+        fill="#a05a2c"
+        stroke="#000000"
+        strokeWidth="1.0"
+        id="rect-rod"
+        width="6"
+        height="48"
+        x="100"
+        y="41"
+        transform="rotate(45 103 69)"
+      />
+      <path
+        fill="#b3b3b3"
+        stroke="#000000"
+        strokeWidth="1.0"
+        d="M117 55 L117 81 L129 81 L129 65 L123 55 Z"
+        id="path-hammer"
+        transform="rotate(-45, 103, 69)"
+      />
+    </>
+  );
+};