+++ web-ui/src/components/common/blob-visuals/blob-parts/tools/Pickaxe.tsx
@@ -1,0 +1,25 @@
+export const Pickaxe = () => {
+  return (
+    <>
+      <rect
+        fill="#a05a2c"
+        stroke="#000000"
+        strokeWidth="1.0"
+        id="rect-rod"
+        width="6"
+        height="48"
+        x="100"
+        y="41"
+        transform="rotate(45 103 69)"
+      />
+      <path
+        fill="#b3b3b3"
+        stroke="#000000"
+        strokeWidth="1.0"
+        d="M 120 49 q 19 20 0 40 7 -20 0 -40 Z"
+        id="path-pickaxe"
+        transform="rotate(-45, 103, 69)"
+      />
+    </>
+  );
+};