+++ web-ui/src/components/common/blob-visuals/utils.tsx
@@ -1,0 +1,21 @@
+import { JSX } from 'react';
+import { ActivityTypeDbo } from '../../../../generated';
+import { Dumbbells } from './blob-parts/tools/Dumbbells';
+import { DoubleDumbbells } from './blob-parts/tools/DoubleDumbbells';
+import { Hammer } from './blob-parts/tools/Hammer';
+import { Pickaxe } from './blob-parts/tools/Pickaxe';
+
+export const mapActivityToTool = (activity?: ActivityTypeDbo | null): JSX.Element | undefined => {
+  switch (activity) {
+    case ActivityTypeDbo.Practice:
+      return <Dumbbells />;
+    case ActivityTypeDbo.IntenseTraining:
+      return <DoubleDumbbells />;
+    case ActivityTypeDbo.Labour:
+      return <Hammer />;
+    case ActivityTypeDbo.Mining:
+      return <Pickaxe />;
+    default:
+      return undefined;
+  }
+};